Gneya

+++ lib/tabscreen.dart
@@ -832,7 +832,7 @@ class _TabScreenState extends State<TabScreen> {
                                   }
                                 }
                                 Map<String,dynamic> product={
-                                  "product_id":int.parse(cart.cartItem[index].productId.toString()),
+                                  "product_id":double.parse(cart.cartItem[index].productId.toString()),
                                   "variation_id":double.parse(variation![index]),
                                   "quantity": cart.cartItem[index].quantity,
                                   "unit_price": cart.cartItem[index].unitPrice,
